chore: 409 - move feature tests to specific folders
