+++ test/feature/login/post_login_test.dart
@@ -11,12 +11,12 @@ import 'package:pass_emploi_app/repositories/crypto/chat_crypto.dart';
 import 'package:pass_emploi_app/repositories/firebase_auth_repository.dart';
 import 'package:redux/src/store.dart';
 
-import '../doubles/dummies.dart';
-import '../doubles/fixtures.dart';
-import '../doubles/stubs.dart';
-import '../redux/middlewares/register_push_notification_token_middleware_test.dart';
-import '../utils/test_setup.dart';
-import 'favoris/offre_emploi_favoris_test.dart';
+import '../../doubles/dummies.dart';
+import '../../doubles/fixtures.dart';
+import '../../doubles/stubs.dart';
+import '../../redux/middlewares/register_push_notification_token_middleware_test.dart';
+import '../../utils/test_setup.dart';
+import '../favoris/offre_emploi_favoris_test.dart';
 
 main() {
   group("after login ...", () {
