+++ test/feature/login/post_logout_test.dart
@@ -4,7 +4,7 @@ import 'package:pass_emploi_app/features/login/login_actions.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
 import 'package:redux/src/store.dart';
 
-import '../utils/test_setup.dart';
+import '../../utils/test_setup.dart';
 
 main() {
   test("After logout user should be signed out from Firebase Auth", () async {
