+++ test/feature/user_action/user_action_create_test.dart
@@ -4,9 +4,9 @@ import 'package:pass_emploi_app/features/user_action/create/user_action_create_s
 import 'package:pass_emploi_app/features/user_action/list/user_action_list_state.dart';
 import 'package:pass_emploi_app/models/user_action.dart';
 
-import '../doubles/fixtures.dart';
-import '../doubles/stubs.dart';
-import '../utils/test_setup.dart';
+import '../../doubles/fixtures.dart';
+import '../../doubles/stubs.dart';
+import '../../utils/test_setup.dart';
 
 main() {
   test("create user action when repo succeeds should display loading and then create user action", () async {
