+++ test/feature/user_action/user_action_delete_test.dart
@@ -6,8 +6,8 @@ import 'package:pass_emploi_app/models/user_action.dart';
 import 'package:pass_emploi_app/models/user_action_creator.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
 
-import '../doubles/stubs.dart';
-import '../utils/test_setup.dart';
+import '../../doubles/stubs.dart';
+import '../../utils/test_setup.dart';
 
 main() {
   test("delete user action when repo succeeds should display loading and then delete user action", () async {
